Cross-validation and testing (all the set of engines)
One important aspect in our cross-validation procedure is that the same engine will not appear in 2 different folds, meaning that a model can see some engine's all the data either during training or validation. For this kind of cross-validation we are using a scikit-learn method GroupKFold, where our groups are engines IDs. An example of cross-validation with sklearn.model_selection.GroupKFold